Make BalanceItem immutable
With mutable BalanceItem there may be issues. When we update an item inside of the repo it also gets updated in the listener too. So there will be no differecen between the updated item and the local item for listener.

To prevent this issue we replace an item with a new update.