Update Zcash sdk 1.3.0 beta16
* Add simplified SDK dependency.

There is no longer a need to map the library to product flavors.

* Add network awareness.

These are the main changes that allow the wallet to use the SDK as one dependency by just providing a network value in the right places.

* Switch to lightwalletd.com server for testnet.

The ECC server is not externally supported and goes down often without notice to external parties. Lightwalletd.com is funded by ZOMG and provides more robust testnet support.

* Remove GlobalScope in favor of synchronizer's scope.

Using GlobalScope usually results in memory leaks and code running when the app is inactive. The Synchronizer can be tied to any scope but it also has it's own lifecycle that ends with 'stop' is called. So that scope can be used in the worst case when there is no over-arching scope to use instead. Typically the activity lifecycle or a subcomponent scope from Dagger would be used, instead. Also, transactions should be accessed after starting the synchronizer to allow things like database migrations to run, first. Although the SDK was recently changed so that this isn't strictly required, it is probably safer for subscription to happen after start, rather than before. Also subscribing during 'init' can make loading the app a little less responsive.

* Fix the send flow.

Previously, send would appear to succeed even when it failed. Also, it would return immediately, rather than waiting for the transaction to be created, which takes about 10 seconds on most devices. The net effect was, after sending it looked like nothing happened for about 10 seconds! Then the pending transaction would appear. Also, the red failure toast behavior would not trigger. Now it does and shows meaningful error messages.

* Update SDK dependency to reflect move to MavenCentral.

* Fix minor typos.

* Show sync percentage.

Right now sync percentage does not show unless lastBlockDate exists. Those two values could probably be decoupled so that percentage can be shown even if there is no date. For now, we can use a hardcoded date on for Zcash, since syncing can take a long time and having a percentage is helpful.